calcFee change size

+++ src/game.cpp
@@ -6610,7 +6610,7 @@ void Game::playerCreateMarketOffer(uint32_t playerId, uint8_t type, uint16_t spr
 		return;
 	}
 
-	uint32_t calcFee = (price / 100.) * amount;
+	uint64_t calcFee = (price / 100.) * amount;
 	uint32_t minFee = std::min<uint32_t>(100000, calcFee);
 	uint32_t fee = std::max<uint32_t>(20, minFee);
